requirements.txt
absl-py==0.9.0
astor==0.8.1
bert-tensorflow==1.0.1
gast==0.3.2
grpcio==1.26.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.18.0
pandas==0.25.3
protobuf==3.11.2
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-hub==0.7.0
termcolor==1.1.0
tqdm==4.41.1
Werkzeug==0.16.0
joblib==0.15.1